+++ packages/core/src/Popup.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo, isValidElement } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
@@ -12,7 +12,6 @@ import PopupButton from './PopupButton';
 
 import Icon from './Icon';
 import Overlay from './Overlay';
-import TextLabel from './TextLabel';
 
 import './styles/_animations.scss';
 import './styles/Popup.scss';
@@ -28,7 +27,9 @@ export const BEM = {
     root: ROOT_BEM,
     container: ROOT_BEM.element('container'),
     body: ROOT_BEM.element('body'),
-    message: ROOT_BEM.element('message'),
+    messageWrapper: ROOT_BEM.element('message-wrapper'),
+    messageTitle: ROOT_BEM.element('message-title'),
+    messageDesc: ROOT_BEM.element('message-desc'),
     button: ROOT_BEM.element('button'),
     buttonsGroup: ROOT_BEM.element('buttons-group')
 };
@@ -40,11 +41,25 @@ PopupIcon.propTypes = {
     type: PropTypes.string.isRequired,
 };
 
-export function PopupMessage({ text }) {
-    return <TextLabel align="center" basic={text} />;
+export function PopupMessage({ title, desc, bottomArea }) {
+    return (
+        <div className={BEM.messageWrapper}>
+            {title && (
+                <span className={BEM.messageTitle}>{title}</span>
+            )}
+            <span className={BEM.messageDesc}>{desc}</span>
+            {bottomArea}
+        </div>
+    );
 }
 PopupMessage.propTypes = {
-    text: PropTypes.string.isRequired,
+    title: PropTypes.string,
+    desc: PropTypes.string.isRequired,
+    bottomArea: PropTypes.node,
+};
+PopupMessage.defaultProps = {
+    title: undefined,
+    bottomArea: undefined,
 };
 
 /**
@@ -73,7 +88,8 @@ function renderPopupButtons(buttons, direction) {
         return (
             <PopupButton
                 key={button.key}
-                {...button.props} />
+                {...button.props}
+            />
         );
     });
 
@@ -90,15 +106,50 @@ function renderPopupButtons(buttons, direction) {
 
 function Popup({
     icon,
+
+    // message area props
+    customMessageNode,
+    messageTitle,
+    messageDesc,
+    messageBottomArea,
+    // message is a legacy prop, should be deprecated in future,
+    //   use `messageDesc` instead for string message,
+    //   use `customMessageNode` instead for node message
     message,
+
+    // button props
     buttons,
     buttonsDirection,
+
     // React props
     className,
     children,
     ...popupProps
 }) {
     const rootClassName = classNames(BEM.root.toString(), className);
+
+    const messageArea = useMemo(
+        () => {
+            if (customMessageNode) {
+                return customMessageNode;
+            }
+
+            // support for legacy node type `message` prop
+            if (message && isValidElement(message)) {
+                return message;
+            }
+
+            return (
+                <PopupMessage
+                    title={messageTitle}
+                    // support for legacy string type `message` prop
+                    desc={messageDesc || message}
+                    bottomArea={messageBottomArea}
+                />
+            );
+        },
+        [message, customMessageNode, messageTitle, messageDesc, messageBottomArea]
+    );
 
     return (
         <div className={rootClassName} {...popupProps}>
@@ -107,7 +158,7 @@ function Popup({
             <div className={BEM.container}>
                 <div className={BEM.body}>
                     {icon && wrapIfNotElement(icon, { with: PopupIcon, via: 'type' })}
-                    {message && wrapIfNotElement(message, { with: PopupMessage, via: 'text' })}
+                    {messageArea}
                 </div>
 
                 {renderPopupButtons(buttons, buttonsDirection)}
@@ -124,14 +175,22 @@ const StringOrElement = PropTypes.oneOfType([
 
 Popup.propTypes = {
     icon: StringOrElement,
+    customMessageNode: PropTypes.node,
+    messageTitle: PropTypes.string,
+    messageDesc: PropTypes.string,
     message: StringOrElement,
+    messageBottomArea: PropTypes.node,
     buttons: PropTypes.arrayOf(PropTypes.element),
     buttonsDirection: PropTypes.oneOf(Object.values(BUTTONS_DIRECTION)),
 };
 
 Popup.defaultProps = {
     icon: null,
+    customMessageNode: undefined,
+    messageTitle: undefined,
+    messageDesc: undefined,
     message: null,
+    messageBottomArea: undefined,
     buttons: [],
     buttonsDirection: BUTTONS_DIRECTION.VERTICAL,
 };